Добавлены CRUD эндпоинты для подписок и проверки на Postman

+++ internal/models/model.go
@@ -6,7 +6,7 @@ type Subscription struct {
 	ID           int        `json:"id"`
 	UserID       string     `json:"user_id"`
 	ServiceName  string     `json:"service_name"`
-	MonthlyPrice int        `json:"monthly_price"`
+	MonthlyPrice float64    `json:"monthly_price"`
 	StartDate    time.Time  `json:"start_date"`
 	EndDate      *time.Time `json:"end_date"`
 }
